Fix a segfault on init under mod_perl2.
In OpenInteract2::Util, _find_descend used a global file handle in a recursive
function, overwriting the handle on each recursion. This caused a segfault
under mod_perl2 when the initialisation was done and destruct was called.
Either using closedir() explicitly in the method, or using lexical file handles
fixes the bug. This commit does the latter.